[Rendering] EffectSystem: Moved CreateEffectCompiler in a separate static class outside of Rendering code
