+++ sources/engine/Xenko.Engine/Engine/Game.cs
@@ -21,6 +21,7 @@ using Xenko.Profiling;
 using Xenko.Rendering;
 using Xenko.Rendering.Fonts;
 using Xenko.Rendering.Sprites;
+using Xenko.Shaders.Compiler;
 using Xenko.Streaming;
 using Xenko.VirtualReality;
 
@@ -370,7 +371,7 @@ namespace Xenko.Engine
             Services.AddService(EffectSystem);
 
             // If requested in game settings, compile effects remotely and/or notify new shader requests
-            EffectSystem.Compiler = EffectSystem.CreateEffectCompiler(Content.FileProvider, EffectSystem, Settings?.PackageName, Settings?.EffectCompilation ?? EffectCompilationMode.Local, Settings?.RecordUsedEffects ?? false);
+            EffectSystem.Compiler = EffectCompilerFactory.CreateEffectCompiler(Content.FileProvider, EffectSystem, Settings?.PackageName, Settings?.EffectCompilation ?? EffectCompilationMode.Local, Settings?.RecordUsedEffects ?? false);
 
             // Setup shader compiler settings from a compilation mode. 
             // TODO: We might want to provide overrides on the GameSettings to specify debug and/or optim level specifically.
